slashCommands/ping.py
import discord
from discord.ext import commands

class Ping(commands.Cog):
    def __init__(self, bot: commands.Bot):
        self.bot = bot

    @discord.app_commands.command(
        name="ping",
        description="Mostra o ping do bot"
    )
    async def ping(self, interaction: discord.Interaction):
        await interaction.response.send_message(
            f"🏓 Pong! {round(self.bot.latency * 1000)}ms"
        )

async def setup(bot: commands.Bot):
    await bot.add_cog(Ping(bot))